improvements

+++ Database/Migrations/2023_07_03_220070_update_mp_payments.php
@@ -14,6 +14,9 @@ return new class extends Migration
     public function up()
     {
         Schema::table('mp_payments', function (Blueprint $table) {
+            if (DB::getDefaultConnection() == 'sqlite') {
+                return;
+            }
             $table->dropForeign('mp_payments_preference_id_foreign');
             $table->dropColumn('preference_id');
         });
